+++ lib/ui/page/home/page/my_profile/microphone_switch/controller.dart
@@ -19,6 +19,7 @@ import 'dart:async';
 
 import 'package:get/get.dart';
 import 'package:medea_jason/medea_jason.dart';
+import 'package:mutex/mutex.dart';
 
 import '/domain/model/media_settings.dart';
 import '/domain/model/ongoing_call.dart';
@@ -32,7 +33,9 @@ export 'view.dart';
 /// Controller of a [MicrophoneSwitchView].
 class MicrophoneSwitchController extends GetxController {
   MicrophoneSwitchController(this._settingsRepository, {String? mic})
-    : _mic = mic ?? _settingsRepository.mediaSettings.value?.audioDevice;
+    : _mic = RxnString(
+        mic ?? _settingsRepository.mediaSettings.value?.audioDevice,
+      );
 
   /// Settings repository updating the [MediaSettings.audioDevice].
   final AbstractSettingsRepository _settingsRepository;
@@ -46,8 +49,11 @@ class MicrophoneSwitchController extends GetxController {
   /// Error message to display, if any.
   final RxnString error = RxnString();
 
+  /// Audio input level of the currently selected microphone.
+  final RxInt level = RxInt(0);
+
   /// ID of the initially selected microphone device.
-  String? _mic;
+  final RxnString _mic;
 
   /// [Worker] reacting on the [MediaSettings] changes updating the [selected].
   Worker? _worker;
@@ -59,17 +65,32 @@ class MicrophoneSwitchController extends GetxController {
   /// [WebUtils.microphonePermission] subscription.
   StreamSubscription? _permissionSubscription;
 
+  /// [LocalMediaTrack] of the currently selected [camera] device.
+  LocalMediaTrack? _localTrack;
+
+  /// [Worker] reacting on the [_mic] changes updating the [_localTrack].
+  Worker? _micWorker;
+
+  /// Mutex guarding [_initTrack].
+  final Mutex _initGuard = Mutex();
+
   @override
   void onInit() async {
+    _micWorker = ever(_mic, (e) => _initTrack());
+
     _devicesSubscription = MediaUtils.onDeviceChange.listen((e) {
       devices.value = e.audio().toList();
-      selected.value = devices.firstWhereOrNull((e) => e.id() == _mic);
+      selected.value = devices.firstWhereOrNull((e) => e.id() == _mic.value);
+
+      if (_mic.value == 'default') {
+        _initTrack();
+      }
     });
 
     _worker = ever(_settingsRepository.mediaSettings, (e) {
       if (e != null) {
-        _mic = e.audioDevice;
-        selected.value = devices.firstWhereOrNull((e) => e.id() == _mic);
+        _mic.value = e.audioDevice;
+        selected.value = devices.firstWhereOrNull((e) => e.id() == _mic.value);
       }
     });
 
@@ -78,7 +99,8 @@ class MicrophoneSwitchController extends GetxController {
       devices.value = await MediaUtils.enumerateDevices(
         MediaDeviceKind.audioInput,
       );
-      selected.value = devices.firstWhereOrNull((e) => e.id() == _mic);
+      selected.value = devices.firstWhereOrNull((e) => e.id() == _mic.value);
+      _initTrack();
     } on UnsupportedError {
       error.value = 'err_media_devices_are_null'.l10n;
     } catch (e) {
@@ -95,14 +117,59 @@ class MicrophoneSwitchController extends GetxController {
 
   @override
   void onClose() {
+    _micWorker?.dispose();
     _devicesSubscription?.cancel();
     _permissionSubscription?.cancel();
     _worker?.dispose();
+    _localTrack?.free();
+    _localTrack = null;
     super.onClose();
   }
 
   /// Sets the provided [device] as a used by default microphone device.
   Future<void> setAudioDevice(DeviceDetails device) async {
     await _settingsRepository.setAudioDevice(device.id());
+  }
+
+  /// Initializes a [RtcVideoRenderer] for the current [_mic].
+  Future<void> _initTrack() async {
+    if (_initGuard.isLocked) {
+      return;
+    }
+
+    level.value = 0;
+    _localTrack?.free();
+    _localTrack = null;
+
+    String? mic = _mic.value;
+
+    await _initGuard.protect(() async {
+      final List<LocalMediaTrack> tracks = await MediaUtils.getTracks(
+        audio: AudioPreferences(
+          device: mic == 'default' ? null : mic,
+          noiseSuppressionLevel:
+              _settingsRepository.mediaSettings.value?.noiseSuppressionLevel,
+          echoCancellation:
+              _settingsRepository.mediaSettings.value?.echoCancellation,
+          autoGainControl:
+              _settingsRepository.mediaSettings.value?.autoGainControl,
+          highPassFilter:
+              _settingsRepository.mediaSettings.value?.highPassFilter,
+        ),
+      );
+
+      if (isClosed) {
+        tracks.firstOrNull?.free();
+        _localTrack = null;
+      } else {
+        _localTrack = tracks.firstOrNull;
+      }
+
+      _localTrack?.onAudioLevelChanged((i) => level.value = i);
+    });
+
+    if (_mic.value != _mic.value && !isClosed) {
+      _initTrack();
+    }
   }
 }