+++ lib/ui/page/home/page/my_profile/microphone_switch/widget/gradient_button.dart
@@ -1,0 +1,75 @@
+// Copyright © 2022-2026 IT ENGINEERING MANAGEMENT INC,
+//                       <https://github.com/team113>
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU Affero General Public License v3.0 as published by the
+// Free Software Foundation, either version 3 of the License, or (at your
+// option) any later version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License v3.0 for
+// more details.
+//
+// You should have received a copy of the GNU Affero General Public License v3.0
+// along with this program. If not, see
+// <https://www.gnu.org/licenses/agpl-3.0.html>.
+
+import 'package:flutter/material.dart';
+
+import '/themes.dart';
+
+/// Rectangular filled with [LinearGradient] selectable button.
+class GradientButton extends StatelessWidget {
+  const GradientButton({super.key, this.label, this.progress = 0});
+
+  /// Label of this [GradientButton].
+  final String? label;
+
+  /// Progress to display [LinearGradient] with.
+  final double progress;
+
+  @override
+  Widget build(BuildContext context) {
+    final style = Theme.of(context).style;
+
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        return Stack(
+          children: [
+            Container(
+              width: double.infinity,
+              decoration: BoxDecoration(
+                color: style.colors.primary,
+                borderRadius: BorderRadius.circular(10),
+              ),
+              child: Padding(
+                padding: const EdgeInsets.all(16.0),
+                child: Text(
+                  label ?? '',
+                  style: style.fonts.normal.regular.onPrimary,
+                  maxLines: 2,
+                  overflow: TextOverflow.ellipsis,
+                ),
+              ),
+            ),
+            AnimatedPositioned(
+              duration: const Duration(milliseconds: 100),
+              curve: Curves.ease,
+              left: 0,
+              top: 0,
+              width: constraints.maxWidth * progress,
+              bottom: 0,
+              child: Container(
+                width: double.infinity,
+                decoration: BoxDecoration(
+                  color: style.colors.onPrimaryOpacity25,
+                ),
+              ),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}